KSENROLL-3145 fixed assembler for intensity and meta data
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-lum/trunk@36911 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-lum-impl/src/test/java/org/kuali/student/lum/program/service/impl/TestProgramServiceImpl.java
@@ -526,7 +526,8 @@ public class TestProgramServiceImpl {
 
             assertNotNull(createdMD.getStdDuration());
             assertEquals("atpDurationTypeKey-test", createdMD.getStdDuration().getAtpDurationTypeKey());
-            assertEquals(new Integer(63), createdMD.getStdDuration().getTimeQuantity());
+// this value is based on the property index (so it may change if properties are added, removed, or re-ordered)
+            assertEquals(new Integer(64), createdMD.getStdDuration().getTimeQuantity());
 
             assertNotNull(createdMD.getStartTerm());
             assertEquals("startTerm-test", createdMD.getStartTerm());
